LPS-169073 Merge the BasePortalCacheManager to EhcachePortalCacheManager
